RoundReport Added: 1.0
Initial release of RoundReport based on plugin.rounds.php from xaseco2.

Made it a bit better with a  GUI of SM-ish like tot tell Scores in Team
mode for now.

Will report a round if players finish, Will also report if nobody
finished.

There might be some bugs as it was tested with one player.

If there are please make a mention in the therefore topic ;)